requirements.txt
httpx
discord.py
git+https://github.com/quarterwire/instapydl.git@master
